server: seperate capacity and leader balancer.
