+++ server/balancer.go
@@ -20,102 +20,25 @@ import (
 	"github.com/pingcap/kvproto/pkg/metapb"
 )
 
+var (
+	_ Balancer = &capacityBalancer{}
+	_ Balancer = &defaultBalancer{}
+	_ Balancer = &leaderBalancer{}
+)
+
 // Balancer is an interface to select store regions for auto-balance.
 type Balancer interface {
 	Balance(cluster *clusterInfo) (*balanceOperator, error)
+	ScoreType() scoreType
 }
 
-var (
-	_ Balancer = &defaultBalancer{}
-	_ Balancer = &resourceBalancer{}
-)
-
-type resourceBalancer struct {
-	balanceFilters        []Filter
-	leaderTransferFilters []Filter
-
-	cfg *BalanceConfig
-}
-
-func newResourceBalancer(cfg *BalanceConfig) *resourceBalancer {
-	rb := &resourceBalancer{cfg: cfg}
-
-	// Add resource filters to check whether the store can only do balance.
-	rb.balanceFilters = append(rb.balanceFilters, newCapacityFilter(cfg))
-	rb.balanceFilters = append(rb.balanceFilters, newSnapCountFilter(cfg))
-
-	// Add leader transfer filters to check whether the store can only do leader transfer.
-	rb.leaderTransferFilters = append(rb.leaderTransferFilters, newLeaderCountFilter(cfg))
-
-	return rb
-}
-
-func (rb *resourceBalancer) filterFromStore(store *storeInfo, filters []Filter, args ...interface{}) bool {
-	for _, filter := range filters {
-		if filter.FilterFromStore(store, args) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (rb *resourceBalancer) filterToStore(store *storeInfo, filters []Filter, args ...interface{}) bool {
-	for _, filter := range filters {
-		if filter.FilterToStore(store, args) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// calculate the score, higher score region will be selected as balance from store,
-// and lower score region will be balance to store. The score range is [0,100].
-// TODO: we should adjust the weight of used ratio and leader score in futher,
-// now it is a little naive.
-func (rb *resourceBalancer) score(store *storeInfo, leaderCount int, regionCount int) int {
-	resourceScorer := newResourceScorer(rb.cfg, leaderCount, regionCount)
-	return resourceScorer.Score(store)
-}
-
-// checkScore checks whether the new store score and old store score are valid.
-func (rb *resourceBalancer) checkScore(cluster *clusterInfo, oldPeer *metapb.Peer, newPeer *metapb.Peer, isLeaderPeer bool) bool {
-	regionCount := cluster.regions.regionCount()
-	oldStore := cluster.getStore(oldPeer.GetStoreId())
-	newStore := cluster.getStore(newPeer.GetStoreId())
-	if oldStore == nil || newStore == nil {
-		log.Debugf("check score failed - old peer: %v, new peer: %v", oldPeer, newPeer)
-		return false
-	}
-
-	// We should check the diff score of pre-balance `from store` and post balance `to store`.
-	// If isLeaderPeer is true, we should calculate the `to store` score with added leader region count.
-	resourceScorer := newResourceScorer(rb.cfg, oldStore.stats.LeaderRegionCount, regionCount)
-	oldStoreScore := resourceScorer.Score(oldStore)
-
-	var newStoreScore int
-	if isLeaderPeer {
-		resourceScorer = newResourceScorer(rb.cfg, newStore.stats.LeaderRegionCount+1, regionCount)
-		newStoreScore = resourceScorer.Score(newStore)
-	} else {
-		resourceScorer = newResourceScorer(rb.cfg, newStore.stats.LeaderRegionCount, regionCount)
-		newStoreScore = resourceScorer.Score(newStore)
-	}
-
-	// If the diff score is in MaxDiffScoreFraction range, then we will do nothing.
-	diffScore := oldStoreScore - newStoreScore
-	if diffScore <= int(float64(oldStoreScore)*rb.cfg.MaxDiffScoreFraction) {
-		log.Debugf("check score failed - diff score is too small - old peer: %v, new peer: %v, old store score: %d, new store score: %d, diif score: %d",
-			oldPeer, newPeer, oldStoreScore, newStoreScore, diffScore)
-		return false
-	}
-
-	return true
-}
-
-func (rb *resourceBalancer) selectFromStore(stores []*storeInfo, excluded map[uint64]struct{}, regionCount int, filters []Filter) *storeInfo {
+func selectFromStore(stores []*storeInfo, excluded map[uint64]struct{}, filters []Filter, st scoreType) *storeInfo {
 	score := 0
+	scorer := newScorer(st)
+	if scorer == nil {
+		return nil
+	}
+
 	var resultStore *storeInfo
 	for _, store := range stores {
 		if store == nil {
@@ -126,11 +49,11 @@ func (rb *resourceBalancer) selectFromStore(stores []*storeInfo, excluded map[ui
 			continue
 		}
 
-		if rb.filterFromStore(store, filters) {
+		if filterFromStore(store, filters) {
 			continue
 		}
 
-		currScore := rb.score(store, store.stats.LeaderRegionCount, regionCount)
+		currScore := scorer.Score(store)
 		if resultStore == nil {
 			resultStore = store
 			score = currScore
@@ -146,8 +69,13 @@ func (rb *resourceBalancer) selectFromStore(stores []*storeInfo, excluded map[ui
 	return resultStore
 }
 
-func (rb *resourceBalancer) selectToStore(stores []*storeInfo, excluded map[uint64]struct{}, regionCount int, filters []Filter) *storeInfo {
+func selectToStore(stores []*storeInfo, excluded map[uint64]struct{}, filters []Filter, st scoreType) *storeInfo {
 	score := 0
+	scorer := newScorer(st)
+	if scorer == nil {
+		return nil
+	}
+
 	var resultStore *storeInfo
 	for _, store := range stores {
 		if store == nil {
@@ -158,11 +86,11 @@ func (rb *resourceBalancer) selectToStore(stores []*storeInfo, excluded map[uint
 			continue
 		}
 
-		if rb.filterToStore(store, filters) {
+		if filterToStore(store, filters) {
 			continue
 		}
 
-		currScore := rb.score(store, store.stats.LeaderRegionCount, regionCount)
+		currScore := scorer.Score(store)
 		if resultStore == nil {
 			resultStore = store
 			score = currScore
@@ -178,10 +106,28 @@ func (rb *resourceBalancer) selectToStore(stores []*storeInfo, excluded map[uint
 	return resultStore
 }
 
+type capacityBalancer struct {
+	filters []Filter
+	st      scoreType
+
+	cfg *BalanceConfig
+}
+
+func newCapacityBalancer(cfg *BalanceConfig) *capacityBalancer {
+	cb := &capacityBalancer{cfg: cfg, st: capacityScore}
+	cb.filters = append(cb.filters, newCapacityFilter(cfg))
+	cb.filters = append(cb.filters, newSnapCountFilter(cfg))
+	return cb
+}
+
+func (cb *capacityBalancer) ScoreType() scoreType {
+	return cb.st
+}
+
 // selectBalanceRegion tries to select a store leader region to do balance and returns true, but if we cannot find any,
 // we try to find a store follower region and returns false.
-func (rb *resourceBalancer) selectBalanceRegion(cluster *clusterInfo, stores []*storeInfo) (*metapb.Region, *metapb.Peer, *metapb.Peer, bool) {
-	store := rb.selectFromStore(stores, cluster.getUnknownStores(), cluster.regions.regionCount(), rb.balanceFilters)
+func (cb *capacityBalancer) selectBalanceRegion(cluster *clusterInfo, stores []*storeInfo) (*metapb.Region, *metapb.Peer, *metapb.Peer, bool) {
+	store := selectFromStore(stores, cluster.getUnknownStores(), cb.filters, cb.st)
 	if store == nil {
 		log.Warn("from store cannot be found to select balance region")
 		return nil, nil, nil, false
@@ -206,43 +152,13 @@ func (rb *resourceBalancer) selectBalanceRegion(cluster *clusterInfo, stores []*
 	return region, leader, nil, true
 }
 
-// selectLeaderTransferRegion tries to select a store leader region to do leader transfer.
-func (rb *resourceBalancer) selectLeaderTransferRegion(cluster *clusterInfo, stores []*storeInfo) (*metapb.Region, *metapb.Peer, *metapb.Peer) {
-	store := rb.selectFromStore(stores, cluster.getUnknownStores(), cluster.regions.regionCount(), rb.leaderTransferFilters)
-	if store == nil {
-		log.Warn("from store cannot be found to select leader transfer region")
-		return nil, nil, nil
-	}
-
-	// Random select one leader region from store.
-	storeID := store.store.GetId()
-	region := cluster.regions.randLeaderRegion(storeID)
-	if region == nil {
-		return nil, nil, nil
-	}
-
-	leader := leaderPeer(region, storeID)
-	if leader == nil {
-		return nil, nil, nil
-	}
-
-	followerPeers, _ := getFollowerPeers(region, leader)
-	newLeader := rb.selectNewLeaderPeer(cluster, followerPeers)
-	if newLeader == nil {
-		log.Warn("new leader peer cannot be found to do leader transfer")
-		return nil, nil, nil
-	}
-
-	return region, leader, newLeader
-}
-
-func (rb *resourceBalancer) selectNewLeaderPeer(cluster *clusterInfo, peers map[uint64]*metapb.Peer) *metapb.Peer {
+func (cb *capacityBalancer) selectNewLeaderPeer(cluster *clusterInfo, peers map[uint64]*metapb.Peer) *metapb.Peer {
 	stores := make([]*storeInfo, 0, len(peers))
 	for storeID := range peers {
 		stores = append(stores, cluster.getStore(storeID))
 	}
 
-	store := rb.selectToStore(stores, nil, cluster.regions.regionCount(), nil)
+	store := selectToStore(stores, cluster.getUnknownStores(), nil, cb.st)
 	if store == nil {
 		log.Warn("find no store to get new leader peer for region")
 		return nil
@@ -252,8 +168,8 @@ func (rb *resourceBalancer) selectNewLeaderPeer(cluster *clusterInfo, peers map[
 	return peers[storeID]
 }
 
-func (rb *resourceBalancer) selectAddPeer(cluster *clusterInfo, stores []*storeInfo, excluded map[uint64]struct{}) (*metapb.Peer, error) {
-	store := rb.selectToStore(stores, excluded, cluster.regions.regionCount(), rb.balanceFilters)
+func (cb *capacityBalancer) selectAddPeer(cluster *clusterInfo, stores []*storeInfo, excluded map[uint64]struct{}) (*metapb.Peer, error) {
+	store := selectToStore(stores, excluded, cb.filters, cb.st)
 	if store == nil {
 		log.Warn("to store cannot be found to add peer")
 		return nil, nil
@@ -272,13 +188,13 @@ func (rb *resourceBalancer) selectAddPeer(cluster *clusterInfo, stores []*storeI
 	return peer, nil
 }
 
-func (rb *resourceBalancer) selectRemovePeer(cluster *clusterInfo, peers map[uint64]*metapb.Peer) (*metapb.Peer, error) {
+func (cb *capacityBalancer) selectRemovePeer(cluster *clusterInfo, peers map[uint64]*metapb.Peer) (*metapb.Peer, error) {
 	stores := make([]*storeInfo, 0, len(peers))
 	for storeID := range peers {
 		stores = append(stores, cluster.getStore(storeID))
 	}
 
-	store := rb.selectFromStore(stores, cluster.getUnknownStores(), cluster.regions.regionCount(), nil)
+	store := selectFromStore(stores, nil, nil, cb.st)
 	if store == nil {
 		log.Warn("from store cannot be found to remove peer")
 		return nil, nil
@@ -288,7 +204,11 @@ func (rb *resourceBalancer) selectRemovePeer(cluster *clusterInfo, peers map[uin
 	return peers[storeID], nil
 }
 
-func (rb *resourceBalancer) doLeaderBalance(cluster *clusterInfo, stores []*storeInfo, region *metapb.Region, leader *metapb.Peer, newPeer *metapb.Peer) (*balanceOperator, error) {
+func (cb *capacityBalancer) doLeaderBalance(cluster *clusterInfo, stores []*storeInfo, region *metapb.Region, leader *metapb.Peer, newPeer *metapb.Peer) (*balanceOperator, error) {
+	if !checkScore(cluster, leader, newPeer, cb.st, cb.cfg) {
+		return nil, nil
+	}
+
 	regionID := region.GetId()
 
 	// If cluster max peer count config is 1, we cannot do leader transfer,
@@ -297,32 +217,25 @@ func (rb *resourceBalancer) doLeaderBalance(cluster *clusterInfo, stores []*stor
 	if meta.GetMaxPeerCount() == 1 {
 		addPeerOperator := newAddPeerOperator(regionID, newPeer)
 		removePeerOperator := newRemovePeerOperator(regionID, leader)
-		if !rb.checkScore(cluster, leader, newPeer, true) {
-			return nil, nil
-		}
 		return newBalanceOperator(region, addPeerOperator, removePeerOperator), nil
 	}
 
-	if !rb.checkScore(cluster, leader, newPeer, false) {
-		return nil, nil
-	}
-
 	followerPeers, _ := getFollowerPeers(region, leader)
-	newLeader := rb.selectNewLeaderPeer(cluster, followerPeers)
+	newLeader := cb.selectNewLeaderPeer(cluster, followerPeers)
 	if newLeader == nil {
 		log.Warn("new leader peer cannot be found to do balance, try to do follower peer balance")
 		return nil, nil
 	}
 
-	leaderTransferOperator := newTransferLeaderOperator(regionID, leader, newLeader, int(rb.cfg.MaxTransferWaitCount))
+	leaderTransferOperator := newTransferLeaderOperator(regionID, leader, newLeader, int(cb.cfg.MaxTransferWaitCount))
 	addPeerOperator := newAddPeerOperator(regionID, newPeer)
 	removePeerOperator := newRemovePeerOperator(regionID, leader)
 
 	return newBalanceOperator(region, leaderTransferOperator, addPeerOperator, removePeerOperator), nil
 }
 
-func (rb *resourceBalancer) doFollowerBalance(cluster *clusterInfo, stores []*storeInfo, region *metapb.Region, follower *metapb.Peer, newPeer *metapb.Peer) (*balanceOperator, error) {
-	if !rb.checkScore(cluster, follower, newPeer, false) {
+func (cb *capacityBalancer) doFollowerBalance(cluster *clusterInfo, stores []*storeInfo, region *metapb.Region, follower *metapb.Peer, newPeer *metapb.Peer) (*balanceOperator, error) {
+	if !checkScore(cluster, follower, newPeer, cb.st, cb.cfg) {
 		return nil, nil
 	}
 
@@ -331,9 +244,9 @@ func (rb *resourceBalancer) doFollowerBalance(cluster *clusterInfo, stores []*st
 	return newBalanceOperator(region, addPeerOperator, removePeerOperator), nil
 }
 
-func (rb *resourceBalancer) doBalance(cluster *clusterInfo) (*balanceOperator, error) {
+func (cb *capacityBalancer) doBalance(cluster *clusterInfo) (*balanceOperator, error) {
 	stores := cluster.getStores()
-	region, leader, follower, isLeaderBalance := rb.selectBalanceRegion(cluster, stores)
+	region, leader, follower, isLeaderBalance := cb.selectBalanceRegion(cluster, stores)
 	if region == nil || leader == nil {
 		log.Warn("region cannot be found to do balance")
 		return nil, nil
@@ -347,9 +260,10 @@ func (rb *resourceBalancer) doBalance(cluster *clusterInfo) (*balanceOperator, e
 	}
 
 	_, excludedStores := getFollowerPeers(region, leader)
+	excludedStores = mergeMap(excludedStores, cluster.getUnknownStores())
 
 	// Select one store to add new peer.
-	newPeer, err := rb.selectAddPeer(cluster, stores, excludedStores)
+	newPeer, err := cb.selectAddPeer(cluster, stores, excludedStores)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
@@ -359,16 +273,95 @@ func (rb *resourceBalancer) doBalance(cluster *clusterInfo) (*balanceOperator, e
 	}
 
 	if isLeaderBalance {
-		ops, err := rb.doLeaderBalance(cluster, stores, region, leader, newPeer)
+		ops, err := cb.doLeaderBalance(cluster, stores, region, leader, newPeer)
 		return ops, errors.Trace(err)
 	}
 
-	return rb.doFollowerBalance(cluster, stores, region, follower, newPeer)
+	return cb.doFollowerBalance(cluster, stores, region, follower, newPeer)
 }
 
-func (rb *resourceBalancer) doLeaderTransfer(cluster *clusterInfo) (*balanceOperator, error) {
+// Balance tries to select a store region to do balance.
+// The priority of balance type is:
+// doBalance:
+// 1 do leader balance.
+// 2 do follower balance.
+func (cb *capacityBalancer) Balance(cluster *clusterInfo) (*balanceOperator, error) {
+	op, err := cb.doBalance(cluster)
+	return op, errors.Trace(err)
+}
+
+type leaderBalancer struct {
+	filters []Filter
+	st      scoreType
+
+	cfg *BalanceConfig
+}
+
+func newLeaderBalancer(cfg *BalanceConfig) *leaderBalancer {
+	lb := &leaderBalancer{cfg: cfg, st: leaderScore}
+	lb.filters = append(lb.filters, newLeaderCountFilter(cfg))
+	return lb
+}
+
+func (lb *leaderBalancer) ScoreType() scoreType {
+	return lb.st
+}
+
+// selectBalanceRegion tries to select a store leader region to do balance.
+func (lb *leaderBalancer) selectBalanceRegion(cluster *clusterInfo, stores []*storeInfo) (*metapb.Region, *metapb.Peer, *metapb.Peer) {
+	store := selectFromStore(stores, cluster.getUnknownStores(), lb.filters, lb.st)
+	if store == nil {
+		log.Warn("from store cannot be found to select balance region")
+		return nil, nil, nil
+	}
+
+	// Random select one leader region from store.
+	storeID := store.store.GetId()
+	region := cluster.regions.randLeaderRegion(storeID)
+	if region == nil {
+		return nil, nil, nil
+	}
+
+	leader := leaderPeer(region, storeID)
+	if leader == nil {
+		return nil, nil, nil
+	}
+
+	followerPeers, _ := getFollowerPeers(region, leader)
+	newLeader := lb.selectNewLeaderPeer(cluster, followerPeers)
+	if newLeader == nil {
+		log.Warn("new leader peer cannot be found to do leader transfer")
+		return nil, nil, nil
+	}
+
+	return region, leader, newLeader
+}
+
+func (lb *leaderBalancer) selectNewLeaderPeer(cluster *clusterInfo, peers map[uint64]*metapb.Peer) *metapb.Peer {
+	stores := make([]*storeInfo, 0, len(peers))
+	for storeID := range peers {
+		stores = append(stores, cluster.getStore(storeID))
+	}
+
+	store := selectToStore(stores, cluster.getUnknownStores(), nil, lb.st)
+	if store == nil {
+		log.Warn("find no store to get new leader peer for region")
+		return nil
+	}
+
+	storeID := store.store.GetId()
+	return peers[storeID]
+}
+
+func (lb *leaderBalancer) doBalance(cluster *clusterInfo) (*balanceOperator, error) {
+	// If cluster max peer count config is 1, we cannot do leader transfer,
+	meta := cluster.getMeta()
+	if meta.GetMaxPeerCount() == 1 {
+		return nil, nil
+	}
+
 	stores := cluster.getStores()
-	region, leader, newLeader := rb.selectLeaderTransferRegion(cluster, stores)
+	region, leader, newLeader := lb.selectBalanceRegion(cluster, stores)
 	if region == nil || leader == nil || newLeader == nil {
 		log.Warn("region cannot be found to do leader transfer")
 		return nil, nil
@@ -381,37 +374,25 @@ func (rb *resourceBalancer) doLeaderTransfer(cluster *clusterInfo) (*balanceOper
 		return nil, nil
 	}
 
-	if !rb.checkScore(cluster, leader, newLeader, true) {
+	if !checkScore(cluster, leader, newLeader, lb.st, lb.cfg) {
 		return nil, nil
 	}
 
 	regionID := region.GetId()
-	leaderTransferOperator := newTransferLeaderOperator(regionID, leader, newLeader, int(rb.cfg.MaxTransferWaitCount))
+	leaderTransferOperator := newTransferLeaderOperator(regionID, leader, newLeader, int(lb.cfg.MaxTransferWaitCount))
 	return newBalanceOperator(region, leaderTransferOperator), nil
 }
 
-// Balance tries to select a store to do balance.
-// The priority of balance type is:
-// doBalance:
-// 1 do leader balance.
-// 2 do follower balance.
-// doLeaderTransfer:
-// 1 do leader transfer.
-func (rb *resourceBalancer) Balance(cluster *clusterInfo) (*balanceOperator, error) {
-	op, err := rb.doBalance(cluster)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if op != nil {
-		return op, nil
-	}
-
-	return rb.doLeaderTransfer(cluster)
+// Balance tries to select a store region to do balance.
+// The balance type is leader transfer.
+func (lb *leaderBalancer) Balance(cluster *clusterInfo) (*balanceOperator, error) {
+	bop, err := lb.doBalance(cluster)
+	return bop, errors.Trace(err)
 }
 
 // defaultBalancer is used for default config change, like add/remove peer.
 type defaultBalancer struct {
-	*resourceBalancer
+	*capacityBalancer
 	region *metapb.Region
 	leader *metapb.Peer
 }
@@ -420,7 +401,7 @@ func newDefaultBalancer(region *metapb.Region, leader *metapb.Peer, cfg *Balance
 	return &defaultBalancer{
 		region:           region,
 		leader:           leader,
-		resourceBalancer: newResourceBalancer(cfg),
+		capacityBalancer: newCapacityBalancer(cfg),
 	}
 }
 
@@ -431,6 +412,8 @@ func (db *defaultBalancer) addPeer(cluster *clusterInfo) (*balanceOperator, erro
 		storeID := peer.GetStoreId()
 		excludedStores[storeID] = struct{}{}
 	}
+
+	excludedStores = mergeMap(excludedStores, cluster.getUnknownStores())
 
 	peer, err := db.selectAddPeer(cluster, stores, excludedStores)
 	if err != nil {
